Update bal_auto_light_dual.weapon
-Decreased range according to Weapons.xlsx sheet.
-Increased **_range_dev to be half of gauss.
-Decreased dmg from 20 to 15
-Changed visual effects to small gauss for better readability, as tiny are barely visible even with full zoom-in on a 32' UHD screen.

Overall, balanced the weapon to be less OP in the tier. Putting it on a strafe bridge becomes less useful in 10vs10+ battles because of the amount of micro needed to aim ships all the time with low maneuverability. However, DEs were not tested against CRs.

Need to update Weapons.xlsx sheet with new data.